Fixed after evaluation fail

+++ libft/libft.h
@@ -15,6 +15,8 @@
 
 # include <unistd.h>
 # include <stdlib.h>
+# include <stdio.h>
+# include <limits.h>
 
 //BONUS
 typedef struct s_list
@@ -54,7 +56,7 @@ int		ft_strncmp(const char *str1, const char *str2, size_t n);
 void	*ft_memchr(const void *str, int c, size_t n);
 int		ft_memcmp(const void *str1, const void *str2, size_t n);
 char	*ft_strnstr(const char *str, const char *to_find, size_t len);
-int		ft_atoi(const char *str);
+long int	ft_atoi(const char *str);
 
 //PART 2
 void	*ft_calloc(size_t count, size_t size);
